refactor gateway env variables
